+++ src/assets/mod.rs
@@ -4,8 +4,8 @@ mod parser;
 pub(crate) mod vector;
 
 pub use asset_loader::VelloAssetLoader;
-pub use vector::Vector;
 pub use vector::VelloAsset;
+pub use vector::VelloAssetData;
 
 pub(crate) use asset_loader::VectorLoaderError;
 pub use parser::*;